+++ src/Commands/Principals/GetSiteGroup.cs
@@ -1,10 +1,8 @@
-﻿using System.Collections.Generic;
-using System.Management.Automation;
+﻿using System.Management.Automation;
 using Microsoft.SharePoint.Client;
-using PnP.PowerShell.Model;
-using PnP.PowerShell.Commands.Base.PipeBinds;
-using Microsoft.Online.SharePoint.TenantAdministration;
+using PnP.PowerShell.Commands.Model;
 using PnP.PowerShell.Commands.Base;
+using PnP.PowerShell.Commands.Base.PipeBinds;
 
 namespace PnP.PowerShell.Commands.Principals
 {
@@ -12,7 +10,7 @@ namespace PnP.PowerShell.Commands.Principals
     public class GetSiteGroup : PnPAdminCmdlet
     {
         [Parameter(Mandatory = false)]
-        public string Site;
+        public SitePipeBind Site;
 
         [Parameter(Mandatory = false)]
         public string Group;
@@ -23,7 +21,7 @@ namespace PnP.PowerShell.Commands.Principals
             var url = PnPConnection.CurrentConnection.Url;
             if (ParameterSpecified(nameof(Site)))
             {
-                url = Site;
+                url = Site.Url;
             }
             var site = this.Tenant.GetSiteByUrl(url);
             if (!ParameterSpecified(nameof(Group)))